+++ app/StreamLink.php
@@ -6,5 +6,5 @@ use Illuminate\Database\Eloquent\Model;
 
 class StreamLink extends Model
 {
-    //
+    // protected $fillable = ["links"];
 }